+++ src/just_makeit/templates/cmake/CMakeLists_module.cmake
@@ -4,7 +4,7 @@ Python3_add_library(<<module>> MODULE WITH_SOABI <<module>>_ext.c)
 target_link_libraries(<<module>> PRIVATE
     <<object_core_libs>>
     <<extra_link_libs_block>>Python3::NumPy)
-target_include_directories(<<module>> PRIVATE ${CMAKE_SOURCE_DIR}/native/inc)
+target_include_directories(<<module>> PRIVATE ${CMAKE_SOURCE_DIR}/native/inc<<extra_include_dirs_block>>)
 if(WIN32 AND CMAKE_C_COMPILER_ID STREQUAL "GNU")
     target_link_options(<<module>> PRIVATE -static-libgcc)
     get_filename_component(gcc_bin "${CMAKE_C_COMPILER}" DIRECTORY)